feat: move overlay UI to the side panel (DAP-4813) (#58)
* Revert "feat: open the overlay by left-clicking on the extension icon (DAP-4805)"

This reverts commit 5470dbdf937837ed14ceb2bef3912054ce389528.

* feat: add skeleton of the side panel (DAP-4813)

* refactor: extract engine from NotificationProvider

* refactor: remove unused imports

* Merge branch 'develop' into dap-4813

* feat: mocked side panel

* refactor: move app switchers to shared components

* refactor: extract engine hooks to a separate package

* refactor: declarative way to initialize parser configs

* feat: synchronize wallet state

* fix: notifications are empty

* refactor: upgrade react router dom

* fix: overlay doesn't work in gateway

* refactor: remove unused imports

* refactor: extract all dependant calls to context

* feat: sync context tree with side panel

* refactor: replace selectedMutation with selectedMutationId

* feat: switch mutation from side panel

* refactor: extract favorite mutations to react-engine

* refactor: extract preferred source to react-engine

* refactor: move mutation editor modal to the side panel

* refactor: rename hook

* refactor: move overlay open state inside

* refactor: move networkId to wallet provider

* refactor: move core inside engine

* refactor: simplify overlay drawer

* feat: remove overlay from extension

* refactor: use wildcard export

* feat: extract notifications to the separate page

* fix: mutation editor not found

* refactor: reuse existing mutation editor

* refactor: get rid of mutations with settings

* fix: broken styles in mutations list

* fix: broken connection between cs and sp

* refactor: memoize mutations

* fix: parser is reattached after mutation switching

* refactor: don't load all mutations in the engine

* refactor: reuse use state hook

* fix: window focus refetch

* fix: empty mutations list

* feat: move apps to the sidepanel

* refactor: replace `/system/` route with `/`

* fix: near social doesn't work

* refactor: extract modalContainerRef into context

* fix: pr review modal doesn't show

* fix: mutation editor doesn't fit into overlay

* refactor: remove unused switchMutationVersion from engine

* refactor: extract setPreferredSource from context

* refactor: remove unused props

* refactor: remove unused mutationApps from the context

* docs: add todos

* feat: keep two engines in sync

* feat: close sidepanel via uber sausage

* fix: broken side panel

* fix: deleted mutation from preferred source is not loaded

* fix: remove from recently used doesn't work

* fix: mutation editor modal is empty

* refactor: extract all engine listeners to one hook

* refactor: return undefined if there is no query data

* fix: sync mutation and apps state

* feat: open wallet in window popup

* feat: exclude wallet urls

* fix: sidepanel reacts on tabs in another windows

* fix: prev is undefined error

* fix: prev is undefined error

* feat: no connection message

* feat: two way state sync

* fix: inconsistent state after forking

* fix: cannot remove mutation from recent

* ci: fix invalid type on

+++ libs/shared-components/src/multitable-panel/components/modals-confirm.tsx
@@ -1,20 +1,24 @@
-import React, { FC, useCallback, useEffect, useMemo, useState } from 'react'
-import styled from 'styled-components'
+import { EntitySourceType, MutationCreateDto, MutationDto } from '@mweb/backend'
 import {
   useCreateMutation,
-  useEditMutation,
-  useMutableWeb,
   useDeleteLocalMutation,
-} from '@mweb/engine'
-import { EntitySourceType, MutationCreateDto, MutationDto } from '@mweb/backend'
-import { Image } from './image'
+  useEditMutation,
+  useMutations,
+  useSetPreferredSource,
+  useSetSelectedMutation,
+  useUpdateMutationLastUsage,
+} from '@mweb/react-engine'
+import React, { FC, useCallback, useEffect, useMemo, useState } from 'react'
+import styled from 'styled-components'
+import { useEngine } from '../../contexts/engine-context'
+import { cloneDeep } from '../../helpers'
 import { useEscape } from '../../hooks/use-escape'
 import { Alert, AlertProps } from './alert'
 import { Button } from './button'
-import { InputImage } from './upload-image'
-import { cloneDeep } from '../../helpers'
-import { DropdownButton } from './dropdown-button'
 import { ButtonsGroup } from './buttons-group'
+import { DropdownButton } from './dropdown-button'
+import { Image } from './image'
+import { InputImage } from './upload-image'
 
 enum MutationModalMode {
   Editing = 'editing',
@@ -243,7 +247,10 @@ export const ModalConfirm: FC<Props> = ({
   const [newDescription, setDescription] = useState<string>(description ?? '')
   const [isApplyToOriginChecked, setIsApplyToOriginChecked] = useState<boolean>(false) // ToDo: separate checkboxes
   const [alert, setAlert] = useState<IAlert | null>(null)
-  const { mutations, switchMutation, switchPreferredSource } = useMutableWeb()
+  const { mutations } = useMutations()
+  const { tree } = useEngine()
+  const { setPreferredSource } = useSetPreferredSource()
+  const { setSelectedMutationId } = useSetSelectedMutation()
 
   const [mode, setMode] = useState(
     !editingMutation.authorId // Newly created local mutation doesn't have author
@@ -261,6 +268,7 @@ export const ModalConfirm: FC<Props> = ({
   const { createMutation, isLoading: isCreating } = useCreateMutation()
   const { editMutation, isLoading: isEditing } = useEditMutation()
   const { deleteLocalMutation } = useDeleteLocalMutation()
+  const { updateMutationLastUsage } = useUpdateMutationLastUsage()
 
   const isFormDisabled = isCreating || isEditing
 
@@ -304,20 +312,26 @@ export const ModalConfirm: FC<Props> = ({
       return
     }
 
+    const hostname = tree?.id
+
+    if (!hostname) throw new Error('No root context ID found')
+
     if (mode === MutationModalMode.Creating || mode === MutationModalMode.Forking) {
       try {
-        const id = await createMutation(
+        const createdMutation = await createMutation(
           mutationToPublish,
           mode === MutationModalMode.Forking
             ? { askOriginToApplyChanges: isApplyToOriginChecked }
             : undefined
         )
-        switchMutation(id)
-        switchPreferredSource(id, EntitySourceType.Origin)
-        await deleteLocalMutation(mutationToPublish.id)
+        setSelectedMutationId(hostname, createdMutation.id)
+        setPreferredSource(createdMutation.id, hostname, EntitySourceType.Origin)
+        updateMutationLastUsage({ mutationId: createdMutation.id, context: tree })
+        deleteLocalMutation(mutationToPublish.id)
         onCloseAll()
       } catch (error: any) {
-        if (error?.message === 'Mutation with that ID already exists') {
+        console.error(error)
+        if (error?.message === 'Item with that ID already exists') {
           setAlert(alerts.idIsNotUnique)
         }
       }
@@ -331,8 +345,8 @@ export const ModalConfirm: FC<Props> = ({
               : { askOriginToApplyChanges: true }
             : undefined
         )
-        switchPreferredSource(mutationToPublish.id, EntitySourceType.Origin)
-        await deleteLocalMutation(mutationToPublish.id)
+        setPreferredSource(mutationToPublish.id, hostname, EntitySourceType.Origin)
+        deleteLocalMutation(mutationToPublish.id)
         onCloseAll()
       } catch (error: any) {
         console.error(error)
